fix backport of 70b19acf
The patch was missing an important hunk of the original diff which
converted another `ln_r` to `systemctl set-default`. Without it, the
initrd would try to boot `multi-user.target` instead of `initrd.target`.
The reason is that `systemctl set-default` modifies `/etc` while `ln_r`
modifies `/usr`. And so the `set-default multi-user.target` in 00systemd
takes priority over the `ln_r initrd.target` in `01systemd-initrd`.

Thanks Jonathan Lebon for spotting this.

Resolves: #1888779